requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
geopy==2.3.0
flask-jwt-extended==4.5.2
requests==2.31.0
python-jose==3.3.0
gunicorn==21.2.0
